upd edit watchlist cache, move to mutations
