+++ src/ducks/Watchlists/Actions/ChangeVisibility/index.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 import Toggle from './Toggle'
-import { useUpdateWatchlist } from '../../gql/hooks'
+import { useUpdateWatchlist } from '../../gql/list/mutations'
 
 const PublicityToggle = ({ watchlist, ...props }) => {
   const [isActive, setActive] = useState(watchlist.isPublic)
