+++ src/ducks/Watchlists/Actions/WeeklyReport/hooks.js
@@ -1,11 +1,13 @@
 import gql from 'graphql-tag'
+import * as Sentry from '@sentry/react'
 import { useMutation } from '@apollo/react-hooks'
-import { updateWatchlistOnEdit } from '../../gql/hooks'
-import { notifyErrorUpdate } from '../../Widgets/TopPanel/notifications'
+import { updateWatchlistOnEdit } from '../../gql/cache'
+import { notifyError } from '../../Widgets/TopPanel/notifications'
 
 const WATCHLIST_MONITORED_MUTATION = gql`
   mutation updateWatchlist($id: Int!, $isMonitored: Boolean) {
     updateWatchlist(id: $id, isMonitored: $isMonitored) {
+      id
       isMonitored
     }
   }
@@ -17,17 +19,12 @@ export const useMonitoringWatchlist = () => {
   })
 
   const updateWatchlist = (id, isMonitored) =>
-    mutate({
-      variables: { isMonitored, id: +id }
-    })
-      .then(
-        ({
-          data: {
-            updateWatchlist: { isMonitored }
-          }
-        }) => isMonitored
-      )
-      .catch(notifyErrorUpdate)
+    mutate({ variables: { isMonitored, id: +id } })
+      .then(({ data: { updateWatchlist } }) => updateWatchlist.isMonitored)
+      .catch(err => {
+        Sentry.captureException(err)
+        notifyError('watchlist', 'update')
+      })
 
   return [updateWatchlist]
 }
