+++ src/ducks/Watchlists/Widgets/TopPanel/BaseActions/index.js
@@ -5,8 +5,8 @@ import ContextMenu from '@santiment-network/ui/ContextMenu'
 import SaveAs from '../../../Actions/SaveAs'
 import { notifyUpdate } from '../notifications'
 import { NonAuthorTrigger, Trigger } from './Items'
-import { useUpdateWatchlist } from '../../../gql/hooks'
 import { useUserWatchlists } from '../../../gql/lists/hooks'
+import { useUpdateWatchlist } from '../../../gql/list/mutations'
 import { getTitleByWatchlistType, SCREENER } from '../../../detector'
 import styles from './index.module.scss'
