+++ src/ducks/Watchlists/gql/hooks.js
@@ -1,59 +1,16 @@
 import { useState, useEffect } from 'react'
-import { useQuery, useMutation } from '@apollo/react-hooks'
+import { useQuery } from '@apollo/react-hooks'
 import { client } from '../../../apollo'
-import {
-  WATCHLIST_SHORT_QUERY,
-  USER_WATCHLISTS_QUERY,
-  UPDATE_WATCHLIST_MUTATION,
-  getRecentWatchlist
-} from './index'
+import { USER_WATCHLISTS_QUERY, getRecentWatchlist } from './index'
 import { PROJECTS_WATCHLIST_QUERY } from '../../../queries/WatchlistGQL'
-import { notifyErrorUpdate } from '../Widgets/TopPanel/notifications'
 import { useUser } from '../../../stores/user'
 import { checkIsNotScreener, stringifyFn } from '../../Screener/utils'
-import { ADDRESS_WATCHLIST_QUERY } from '../../WatchlistAddressesTable/gql/queries'
-import { BLOCKCHAIN_ADDRESS } from '../detector'
 
 const EMPTY_ARRAY = []
 const DEFAULT_WATCHLISTS = []
 
-function buildWatchlistCacheUpdater (reducer) {
-  return (cache, { data }) => {
-    const query =
-      data.updateWatchlist.type === BLOCKCHAIN_ADDRESS
-        ? ADDRESS_WATCHLIST_QUERY
-        : PROJECTS_WATCHLIST_QUERY
-
-    const watchlist = cache.readQuery({
-      query: query,
-      variables: { id: +data.updateWatchlist.id }
-    })
-
-    cache.writeQuery({
-      query: query,
-      variables: { id: +data.updateWatchlist.id },
-      data: { watchlist: reducer(data, watchlist) }
-    })
-  }
-}
-
-export const updateWatchlistOnEdit = buildWatchlistCacheUpdater(
-  ({ updateWatchlist }, watchlist) => ({ ...watchlist, ...updateWatchlist })
-)
-
 export function useWatchlist ({ id, skip }) {
   const { data, loading, error } = useQuery(PROJECTS_WATCHLIST_QUERY, {
-    skip: !id || skip,
-    variables: {
-      id: +id
-    }
-  })
-
-  return [data ? data.watchlist : undefined, loading, error]
-}
-
-export function useShortWatchlist ({ id, skip }) {
-  const { data, loading, error } = useQuery(WATCHLIST_SHORT_QUERY, {
     skip: !id || skip,
     variables: {
       id: +id
@@ -120,31 +77,6 @@ export function useRecentWatchlists (watchlistsIDs) {
   )
 
   return [recentWatchlists, isLoading, isError]
-}
-
-export function useUpdateWatchlist () {
-  const [mutate, data] = useMutation(UPDATE_WATCHLIST_MUTATION, {
-    update: updateWatchlistOnEdit
-  })
-
-  function updateWatchlist (watchlist, newParams) {
-    const { id, isPublic, name, description, function: oldFn } = watchlist
-
-    return mutate({
-      variables: {
-        id: +id,
-        isPublic:
-          newParams.isPublic === undefined ? isPublic : newParams.isPublic,
-        name: newParams.name || name,
-        description: newParams.description || description,
-        function: stringifyFn(newParams.function || oldFn)
-      }
-    })
-      .then(({ data }) => ({ ...watchlist, ...data.updateWatchlist }))
-      .catch(notifyErrorUpdate)
-  }
-
-  return [updateWatchlist, data]
 }
 
 export function getProjectsByFunction (func, query) {
